requirements.txt
py-cord 
pymongo
qrcode
discord_webhook
pillow
matplotlib
python-dotenv
colorama
requests
